pyproject.toml
[project]
name = "tds2-backup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "beautifulsoup4==4.13.4",
    "certifi==2025.8.3",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "contourpy==1.3.3",
    "cycler==0.12.1",
    "distro==1.9.0",
    "duckdb==1.3.2",
    "fastapi==0.116.1",
    "fonttools==4.59.0",
    "groq==0.31.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "idna==3.10",
    "kiwisolver==1.4.8",
    "lxml==6.0.0",
    "matplotlib==3.10.5",
    "numpy==2.3.2",
    "packaging==25.0",
    "pandas==2.3.1",
    "pillow==11.3.0",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "requests==2.32.4",
    "seaborn==0.13.2",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.7",
    "starlette==0.47.2",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "uvloop==0.21.0",
    "watchfiles==1.1.0",
    "websockets==15.0.1",
]
